+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
  - OPSWAT anti-virus integration
+ - Fix health check
  - Import publishers from sdc.rabbit module, not sdc.rabbit.publisher
 
 ### 1.4.0 2018-02-05